Extract tests for GroupNodes rewriter
